Make the plugin example completely independent
