+++ Dockerfile
@@ -4,4 +4,4 @@ RUN mkdir server
 RUN mkdir plugin
 COPY ./server/target/cqf-ruler-server-*.war server/ROOT.war
 EXPOSE 8080
-CMD ["java", "-cp", "server/ROOT.war", "-Dloader.path=WEB-INF/classes,WEB-INF/lib,WEB-INF/lib-provided$,plugin", "org.springframework.boot.loader.PropertiesLauncher"]
+CMD ["java", "-cp", "server/ROOT.war", "-Dloader.path=WEB-INF/classes,WEB-INF/lib,WEB-INF/lib-provided,plugin", "org.springframework.boot.loader.PropertiesLauncher"]
