+++ example/Dockerfile
@@ -1,0 +1,3 @@
+FROM contentgroup/cqf-ruler:latest
+
+COPY ./target/cqf-ruler-*.jar plugin
