+++ example/src/main/resources/application.yaml
