+++ external/README.MD
@@ -2,7 +2,7 @@
 
 ## Overview
 
-This project only exists to copy and compile the sources from the [HAPI FHIR JPA Server Starter](https://github.com/hapifhir/hapi-fhir-jpaserver-starter) project. The cqf-ruler builds on the starter server, but class jars are not regularly published for that project so we create a `cqf-ruler-external` library here.
+This project only exists to copy and compile the sources from the [HAPI FHIR JPA Server Starter](https://github.com/hapifhir/hapi-fhir-jpaserver-starter) project. The cqf-ruler builds on the starter server, but class jars are not regularly published for that project so we create a `cqf-ruler-external` library here. If the starter server becomes regularly published, this module will be retired.
 
 The `hapi-fhir-jpaserver-starter` project is added as a git sub-module. It's pinned to specific tag. From there, the src code is copied as a `generated-sources` for the `cql-ruler-external` project and minimally processed to suppress warnings and change the package name so that we don't publish anything that would collide with Smile's work.
 
@@ -31,6 +31,6 @@ Next, attempt to build the project. All the dependency versions in the `cql-rule
 
 ## Future Work
 
-Architecturally, we'd like the cqf-ruler components to be able to be dropped into the `hapi-fhir-server` directly. There are a few dependencies that exist in the cqf-ruler projects that make that not possible for the time being. We need to coordinate with the Smile/HAPI Team to identify of subset of this code that could be migrated to the HAPI base server. For example:
+Eventually, it'd be ideal for the cqf-ruler components to be able to be dropped into the `hapi-fhir-server` directly. There are a few dependencies that exist in the cqf-ruler projects that make that not possible for the time being. We need to coordinate with the Smile/HAPI Team to identify of subset of this code that could be migrated to the HAPI base server. For example:
 
 * Condition Annotations (e.g. `OnDstuCondition.class`)
