--- plugin/hello-world/.settings/org.eclipse.jdt.apt.core.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.jdt.apt.aptEnabled=false
