--- plugin/hello-world/pom.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.opencds.cqf.ruler</groupId>
-		<artifactId>cqf-ruler-plugin</artifactId>
-		<version>0.5.0-SNAPSHOT</version>
-	</parent>
-
-	<artifactId>cqf-ruler-hello-world</artifactId>
-</project>
